Fri, Aug 2, 2024, 1:11 AM -05:00

+++ readability/readability.c
@@ -22,8 +22,8 @@ int main(void)
 // Compute Coleman-Liau formula
 // use round function from math.h
 // cast as float so decimals arent truncated.
-    float L = (float) letters / (float) words * 100;
-    float S = (float) sentences / (float) words * 100;
+    float L = (float) letters / words * 100;
+    float S = (float) sentences / words * 100;
     float index = 0.0588 * L - 0.296 * S - 15.8;
 
     int grade_level = round(index);
